requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
soupsieve==2.7
tba-api-v3client @ git+https://github.com/TBA-API/tba-api-client-python.git@4f6ded8fb4bf8f7896891a9aa778ce15a2ef720b
typing_extensions==4.14.0
urllib3==2.4.0
